+++ custom_components/navet/frontend/assets/vendor-8vOrY_Mc.js
@@ -1,0 +1,1 @@
+var e=`/api/navet/static/assets/dashboard-widgets-vh-t_kPv.css`;export{e as t};