+++ src/types/navet-runtime-config.d.ts
@@ -2,6 +2,7 @@ interface NavetRuntimeConfig {
   hassUrl?: string;
   hassToken?: string;
   dashboardConfigUrl?: string;
+  proxyBaseUrl?: string;
 }
 
 interface Window {